examples: fix sha256-async hang in release mode
It turns out that we had a race condition between the time we read
STATUS register and the DMA actually starting the transfer. If we
modify the driver such that we use HASHCRYPT interrupts for errors and
digest ready, we avoid the race condition altogether and the example
works in both dev and release modes.

Note that, ideally, `transfer()` would produce a `Result` to indicate
the possible error condition. Because that would result in a breaking
change, this PR is only fixing the bug and assuming the error
condition never triggers, which is no different from what the driver
was already doing anyway.

Fixes #248

Signed-off-by: Felipe Balbi <febalbi@microsoft.com>