i2c: master: filter to frequencies strictly less than bus speed
Disallow configurations that would produce a nominal frequency equal
to requested bus frequency. The reason for this is that all clocks
used as I2C interface clock have some error margin in them and if they
were to generate a frequency slightly higher than expected, we would
end up with an out-of-spec frequency. To avoid that, generate a
frequency slightly below target in hopes that errors will never cross
that threshold.

Signed-off-by: Felipe Balbi <febalbi@microsoft.com>