Major rework of the clocking subsystem
This is a squashed version of the `james/clock-reimpl-reorg` branch.

This change is part of a major rework to move the embassy-imxrt
project into using a clocking pattern consistent with other HALs
like embassy-stm32.

This change centralizes the clock configuration into two major stages:

* An "init" stage that configures the system clocks in a static
  fashion, not changeable after initialization
* A "runtime" stage, where individual peripherals are configured and
  enabled or disabled

This change exposes more options for clocking, and enforces that
the selected configuration is coherent: e.g. that all clocks that
are selected are actually enabled and within a reasonable range.