Audit timers HAL #3: Remove unwraps (#532)
Depends on #531 
Resolves #530 

This removes the last of the panic paths in the timers HAL by removing
unwraps and changing a few constructors to return a `Result` instead
(which introduces a breaking change). A new freq field was added to the
PWM struct since originally, we unwrapped in a PWM trait method which
does not return a `Result`. So, we just verify the clock config is valid
in the constructor and can use the cached freq in the trait methods.